[e2e] A11y Improve e2e to avoid flakiness (#2474)
* improve a11y to reduce flakiness
